data_lake_etl/pyproject.toml
[tool.poetry]
name = "data_lake_etl"
version = "0.1.0"
description = ""
authors = ["kbpark-onepredict"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
rich = "^10.16.1"
pandas = "^1.3.5"
dash = "^2.0.0"
pymongo = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"
mypy = "^0.930"
wemake-python-styleguide = "^0.16.0"
ipykernel = "^6.6.0"
MonkeyType = "^21.5.0"
blackcellmagic = "^0.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
